test(portal-next): split tests into 3 different files
To manipulate file upload, this commit add a dev dependency on @testing-library/angular. It helps to handle file change event.
