+++ gravitee-apim-portal-webui-next/package.json
@@ -76,6 +76,7 @@
     "@storybook/angular": "8.0.5",
     "@storybook/blocks": "8.0.5",
     "@storybook/test": "8.0.5",
+    "@testing-library/angular": "^17.2.1",
     "@types/jest": "29.5.12",
     "@types/js-yaml": "4.0.9",
     "@typescript-eslint/eslint-plugin": "7.2.0",
